save breakpoint change to db
