+++ app/shared/db/breakpoints.server.ts
@@ -1,8 +1,11 @@
 import {
   initialBreakpoints,
-  type Project,
   defaultBreakpoint,
+  type Project,
+  type Tree,
+  type Breakpoint,
 } from "@webstudio-is/sdk";
+import { applyPatches, type Patch } from "immer";
 import { prisma } from "./prisma.server";
 
 export const create = async (projectId: Project["id"]) => {
@@ -13,6 +16,7 @@ export const create = async (projectId: Project["id"]) => {
       projectId,
     }));
   // https://github.com/prisma/prisma/issues/8131
+  // @todo run this in parallel
   const breakpoints = await prisma.$transaction(
     data.map((breakpoint) => prisma.breakpoint.create({ data: breakpoint }))
   );
@@ -28,4 +32,26 @@ export const load = async (projectId: Project["id"]) => {
   // DB never contains the default breakpoint
   breakpoints.push(defaultBreakpoint);
   return breakpoints;
+};
+
+export const patch = async (
+  { projectId }: { treeId: Tree["id"]; projectId: Project["id"] },
+  patches: Array<Patch>
+) => {
+  const breakpoints = await await prisma.breakpoint.findMany({
+    where: { projectId },
+  });
+
+  const nextBreakpoints = applyPatches<Array<Breakpoint>>(breakpoints, patches);
+
+  await Promise.all(
+    Object.values(nextBreakpoints).map(async (breakpoint) => {
+      const { id, ...rest } = breakpoint;
+      await prisma.breakpoint.upsert({
+        where: { id },
+        create: breakpoint,
+        update: rest,
+      });
+    })
+  );
 };
